Experiments with blitting
The pico doesn't like doing an LDIR instruction from ROM to
screen so have experimented here.

This example keeps the CSROM on once a ROM read has been
served and this appears to fix corruption issues.

More work is required but checking this in on a branch as a checkout.