Attempt to keep /CSROM for the duration
Previously I was toggling the /CSROM each time a byte was supplied by the pico.
This worked fine on 2 of my spectrums but not on the 3rd. So changed the code
so put data pio pulls two values, one of the byte to write, the next is whether to
keep /CSROM high.